Refactor filtering to repository
