+++ app/src/main/java/com/cornellappdev/score/model/Game.kt
@@ -278,12 +278,3 @@ fun List<GameDetailsBoxScore>.toScoreEvents(teamLogo: String): List<ScoreEvent> 
         )
     }
 }
-
-val validSports = setOf(
-    "Baseball", "Basketball", "Field Hockey",
-    "Football", "Ice Hockey", "Lacrosse", "Soccer"
-)
-
-fun isValidSport(sportName: String): Boolean {
-    return sportName in validSports
-}
