+++ app/src/main/java/com/cornellappdev/score/model/ScoreRepository.kt
@@ -12,6 +12,7 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 import javax.inject.Singleton
+import com.cornellappdev.score.util.isValidSport
 
 /**
  * This is a singleton responsible for fetching and caching all data for Score.
@@ -45,34 +46,35 @@ class ScoreRepository @Inject constructor(
                 val games = result.getOrNull()
 
                 val gamesList: List<Game> =
-                    games?.games?.mapNotNull { game ->
-                        /**
-                         * The final scores in the past game cards are obtained by parsing a String
-                         * result from the GameQuery, which is oftentimes in the format
-                         * Result, CornellScore-OpponentScore (e.g. "W, 2-1"). Not all of the strings
-                         * are in this format (e.g. 4th of 6, 1498 points for women's Swimming and
-                         * Diving), but in this case, the cornellScore and otherScore parameters of
-                         * the game and associated card should be null, and as of right now,
-                         * null-scored games are filtered out.
-                         */
-                        val scores = game?.result?.split(",")?.getOrNull(1)?.split("-")
-                        val cornellScore = scores?.getOrNull(0)?.toNumberOrNull()
-                        val otherScore = scores?.getOrNull(1)?.toNumberOrNull()
-                        game?.team?.image?.let {
-                            Game(
-                                id = game.id ?: "", // Should never be null
-                                teamLogo = it,
-                                teamName = game.team.name,
-                                teamColor = parseColor(game.team.color).copy(alpha = 0.4f * 255),
-                                gender = if (game.gender == "Mens") "Men's" else "Women's",
-                                sport = game.sport,
-                                date = game.date,
-                                city = game.city,
-                                cornellScore = cornellScore,
-                                otherScore = otherScore
-                            )
-                        }
-                    } ?: emptyList()
+                    games?.games?.filter { game -> isValidSport(game?.sport ?: "") }
+                        ?.mapNotNull { game ->
+                            /**
+                             * The final scores in the past game cards are obtained by parsing a String
+                             * result from the GameQuery, which is oftentimes in the format
+                             * Result, CornellScore-OpponentScore (e.g. "W, 2-1"). Not all of the strings
+                             * are in this format (e.g. 4th of 6, 1498 points for women's Swimming and
+                             * Diving), but in this case, the cornellScore and otherScore parameters of
+                             * the game and associated card should be null, and as of right now,
+                             * null-scored games are filtered out.
+                             */
+                            val scores = game?.result?.split(",")?.getOrNull(1)?.split("-")
+                            val cornellScore = scores?.getOrNull(0)?.toNumberOrNull()
+                            val otherScore = scores?.getOrNull(1)?.toNumberOrNull()
+                            game?.team?.image?.let {
+                                Game(
+                                    id = game.id ?: "", // Should never be null
+                                    teamLogo = it,
+                                    teamName = game.team.name,
+                                    teamColor = parseColor(game.team.color).copy(alpha = 0.4f * 255),
+                                    gender = if (game.gender == "Mens") "Men's" else "Women's",
+                                    sport = game.sport,
+                                    date = game.date,
+                                    city = game.city,
+                                    cornellScore = cornellScore,
+                                    otherScore = otherScore
+                                )
+                            }
+                        } ?: emptyList()
                 _upcomingGamesFlow.value = ApiResponse.Success(gamesList)
             } else {
                 _upcomingGamesFlow.value = ApiResponse.Error
